requirements.txt
python-jobspy==1.1.80
anthropic>=0.25.0
gspread>=6.0.0
google-auth>=2.28.0
apscheduler>=3.10.4
python-dotenv>=1.0.1
pandas>=2.2.0
loguru>=0.7.2
